backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "beautifulsoup4>=4.14.3",
    "dnspython>=2.8.0",
    "email-validator>=2.3.0",
    "fastapi>=0.135.3",
    "httpx>=0.28.1",
    "openai>=2.31.0",
    "openai-agents>=0.13.6",
    "pandas>=3.0.2",
    "passlib[bcrypt]>=1.7.4",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "serpapi>=1.0.2",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.44.0",
]
